Issue #166 Download UTXO set

+++ ZcashLightClientKit/Entity/UnspentTransactionOutputEntity.swift
@@ -11,14 +11,14 @@ public protocol UnspentTransactionOutputEntity {
     
     var address: String { get set }
     
-    var txid: Data {get set}
+    var txid: Data { get set }
 
-    var index: Int32 {get set}
+    var index: Int { get set }
 
-    var script: Data {get set}
+    var script: Data { get set }
 
-    var valueZat: Int64 {get set}
+    var valueZat: Int { get set }
 
-    var height: UInt64 {get set}
+    var height: Int { get set }
 
 }
